+++ docs/CHANGELOG.md
@@ -16,6 +16,8 @@ Entries inside each section should be ordered by type:
 !-->
 # unreleased - YYYY-MM-DD
 ## Python API
+* [Added] `quilt3.search()` and `quilt3.Bucket.search()` now accept custom Elasticsearch queries ([#3448](https://github.com/quiltdata/quilt/pull/3448))
+* [Fixed] `quilt3.search()` and `quilt3.Bucket.search()` now work with 2022+ Quilt stacks ([#3448](https://github.com/quiltdata/quilt/pull/3448))
 
 ## CLI